+++ src/main/java/kr/startoff/backend/domain/user/domain/oauth2/OAuth2UserInfoFactory.java
@@ -1,4 +1,4 @@
-package kr.startoff.backend.global.security.oauth2.user;
+package kr.startoff.backend.domain.user.domain.oauth2;
 
 import static kr.startoff.backend.global.exception.ExceptionType.*;
 
@@ -6,7 +6,6 @@ import java.util.Map;
 
 import kr.startoff.backend.domain.user.domain.AuthProvider;
 import kr.startoff.backend.domain.user.exception.UserException;
-import kr.startoff.backend.global.exception.ExceptionType;
 
 public class OAuth2UserInfoFactory {
 	public static OAuth2UserInfo getOAuth2UserInfo(String registrationId, Map<String, Object> attributes) {
