+++ src/main/java/kr/startoff/backend/global/config/SecurityConfig.java
@@ -17,8 +17,8 @@ import kr.startoff.backend.global.security.jwt.AuthTokenFilter;
 import kr.startoff.backend.global.security.jwt.JwtAccessDeniedHandler;
 import kr.startoff.backend.global.security.jwt.JwtAuthEntryPoint;
 import kr.startoff.backend.global.security.jwt.JwtUtil;
-import kr.startoff.backend.global.security.oauth2.CustomOAuth2UserService;
-import kr.startoff.backend.global.security.oauth2.HttpCookieOAuth2AuthorizationRequestRepository;
+import kr.startoff.backend.domain.user.service.CustomOAuth2UserService;
+import kr.startoff.backend.domain.user.repository.HttpCookieOAuth2AuthorizationRequestRepository;
 import kr.startoff.backend.global.security.oauth2.OAuth2AuthenticationFailureHandler;
 import kr.startoff.backend.global.security.oauth2.OAuth2AuthenticationSuccessHandler;
 import kr.startoff.backend.domain.user.service.UserDetailsServiceImpl;
